feat(Aspect): unify attributes, properties, listeners into a new enum
This reduces the size of the Element struct (and therefore the Node enum)
by 6 words as there are 2 less `Vec`s in it now.

The Jfb benchmark now runs approximately 10% faster, from a 1.66 slowdown
compared to vanillajs (keyed), to 1.50. The overhead from raw wasm-bindgen is
about 40%, which itself has a slowdown of 1.07 from vanillajs.

The improvement in non-keyed is also similar -- about 10%, from 1.44 to 1.30.